requirements.txt
google-generativeai
langchain
langchain-google-genai
pandas
matplotlib
streamlit
python-dotenv
pytest
